+++ lib/praxis/action_definition.rb
@@ -73,10 +73,6 @@ module Praxis
       end
       self.instance_eval(&ApiDefinition.instance.traits[trait_name])
     end
-
-    def canonical_path
-      resource_definition.canonical_path_action self.name
-    end
     
     def params(type=Attributor::Struct, **opts, &block)
       return @params if !block && type == Attributor::Struct